Move prepare_load_package to tests/load/utils
