+++ tests/load/test_dummy_client.py
@@ -22,6 +22,7 @@ from dlt.destinations.dummy.configuration import DummyClientConfiguration
 from dlt.load.exceptions import LoadClientJobFailed, LoadClientJobRetry
 
 from tests.utils import clean_test_storage, init_test_logging, TEST_DICT_CONFIG_PROVIDER, preserve_environ
+from tests.load.utils import prepare_load_package
 
 
 NORMALIZED_FILES = [
@@ -390,20 +391,6 @@ def run_all(load: Load) -> None:
         if metrics.pending_items == 0:
             return
         sleep(0.1)
-
-
-def prepare_load_package(load_storage: LoadStorage, cases: Sequence[str]) -> Tuple[str, Schema]:
-    load_id = uniq_id()
-    load_storage.create_temp_load_package(load_id)
-    for case in cases:
-        path = f"./tests/load/cases/loading/{case}"
-        shutil.copy(path, load_storage.storage.make_full_path(f"{load_id}/{LoadStorage.NEW_JOBS_FOLDER}"))
-    for f in ["schema_updates.json", "schema.json"]:
-        path = f"./tests/load/cases/loading/{f}"
-        shutil.copy(path, load_storage.storage.make_full_path(load_id))
-    load_storage.commit_temp_load_package(load_id)
-    schema = load_storage.load_package_schema(load_id)
-    return load_id, schema
 
 
 def setup_loader(delete_completed_jobs: bool = False, client_config: DummyClientConfiguration = None) -> Load:
